requirements.txt
tdqm
Pillow